ISIS-2711: merges LazyTreeNode up into TreeNode
because:
* LazyTreeNode is the only implementation of TreeNode
* value types should not be part of an inheritance hierarchy
